requirements.txt
altgraph==0.17
certifi==2021.5.30
charset-normalizer==2.0.1
colorama==0.4.4
decorator==4.4.2
ffmpeg-python==0.2.0
future==0.18.2
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.4
numpy==1.21.0
pefile==2021.5.24
Pillow==8.3.1
proglog==0.1.9
pyinstaller==4.4
pyinstaller-hooks-contrib==2021.2
PySimpleGUI==4.45.0
pytube==10.9.2
pywin32-ctypes==0.2.0
requests==2.26.0
tqdm==4.61.2
urllib3==1.26.6
